Created FirestoreService class, moved "save" function definition from MVM class to new class.
