+++ app/src/main/java/edu/ucandroid/weathernotice/ui/main/MainViewModel.kt
@@ -6,24 +6,19 @@ import androidx.lifecycle.ViewModel
 import com.google.firebase.auth.FirebaseUser
 import edu.ucandroid.weathernotice.dto.Weather
 import edu.ucandroid.weathernotice.service.WeatherService
-import com.google.firebase.firestore.FirebaseFirestore
 import edu.ucandroid.weathernotice.dto.LocationInfo
 import edu.ucandroid.weathernotice.dto.Reminder
+import edu.ucandroid.weathernotice.service.FirestoreService
 import edu.ucandroid.weathernotice.service.LocationService
 
 class MainViewModel : ViewModel() {
 
-    private lateinit var firestore : FirebaseFirestore
+    var firestore: FirestoreService = FirestoreService()
     var locationinfos: MutableLiveData<ArrayList<LocationInfo>> = MutableLiveData()
     var locationService: LocationService = LocationService()
     var weatherLocations: MutableLiveData<ArrayList<Weather>> = MutableLiveData<ArrayList<Weather>>()
     var weatherService: WeatherService = WeatherService()
 
-    init{
-        fetchLocations()
-        fetchWeatherLocations("e")
-        firestore = FirebaseFirestore.getInstance()
-    }
 
 
     fun fetchLocations() {
@@ -41,22 +36,8 @@ class MainViewModel : ViewModel() {
     fun fetchWeatherLocations(locationName: String) {
         weatherLocations = weatherService.fetchWeatherByLocation(locationName)
     }
-    fun save(reminder: Reminder, user: FirebaseUser){
-        val document = firestore.collection("weather").document()
-        reminder.city = document.id
-        val set = document.set(reminder)
-        set.addOnSuccessListener {
-            Log.d("Firebase", "document saved")
-           /**
-            if (reminder != null && reminder.size > 0) {
-                saveReminders(reminder, user)
-            }
-            */
-        }
-        set.addOnFailureListener {
-            Log.d("Firebase", "Save Failed")
-        }
-
+    fun save(reminder: Reminder, user: FirebaseUser) {
+        firestore.save(reminder, user)
     }
 
 }
